fix: Correcting multi-vibe behaviour
Looks like the single index send commands got commented out during the
ButtplugClientDevice refactor.

This change simply reintroduces the manual command construction for
per-vibe sliders. It could also be done via Vuex or similar to manage
the array of speeds, but that seemed overkill for something that's not
a big deal.

Fixes #7